+++ RELEASES.md
@@ -3,7 +3,8 @@
 ## 0.9.1dev
 
 #### New features
-
+- Make alpha parameter in Fused Gromov Wasserstein differentiable (PR #463)
+- Added the sparsity-constrained OT solver to `ot.smooth` and added ` projection_sparse_simplex` to `ot.utils` (PR #459)
 #### Closed issues
 
 - Fix circleci-redirector action and codecov (PR #460)